修改goods分页方法，转换日期格式
