+++ src/main/java/com/book/mall/mall/mapper/GoodsMapper.java
@@ -18,7 +18,6 @@ public interface GoodsMapper {
             "<if test='name != null'> AND name like #{name} </if>" +
             "<if test='kind != null'> AND kind = #{kind} </if>" +
             "</where>" +
-            "limit #{start} offset #{end} " +
             "</script>")
     public Long getTotal(GoodsFindReqForm reqForm);
