+++ src/main/java/com/book/mall/mall/reqform/GoodsFindReqForm.java
@@ -1,46 +1,10 @@
 package com.book.mall.mall.reqform;
 
-public class GoodsFindReqForm {
+public class GoodsFindReqForm extends BaseReqForm{
 
     private Long id;
     private String name;
     private String kind;
-    private Integer pageSize = 1000;
-    private Integer pageNo = 1;
-    private Integer start = 1000;
-    private Integer end = 1;
-
-    public Integer getStart() {
-        return start;
-    }
-
-    public void setStart(Integer start) {
-        this.start = start;
-    }
-
-    public Integer getEnd() {
-        return end;
-    }
-
-    public void setEnd(Integer end) {
-        this.end = end;
-    }
-
-    public Integer getPageSize() {
-        return pageSize;
-    }
-
-    public void setPageSize(Integer pageSize) {
-        this.pageSize = pageSize;
-    }
-
-    public Integer getPageNo() {
-        return pageNo;
-    }
-
-    public void setPageNo(Integer pageNo) {
-        this.pageNo = pageNo;
-    }
 
     public Long getId() {
         return id;
